Refactor code by removing and adding informative print statements
This commit includes a removal of a print statement that informed about the new height after a scroll operation in the main script. In addition, a more consequential print statement is now included in the code to show the exact number of new businesses added after reading the updated data. The way businesses are obtained has also been simplified, removing unused variables.